feature: implemented utils class, organizing some methods and free functions we already implemented
