+++ include/Response.hpp
@@ -67,4 +67,3 @@ public:
 };
 
 std::ostream &operator<<(std::ostream &out, const Response &obj);
-std::string intToString(int value);
